Services/mvpgpt/requirements.txt
fastapi
uvicorn
httpx
websockets
google-genai